requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
docstring_parser==0.16
fonttools==4.57.0
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instructor==1.7.9
iso8601==2.1.0
Jinja2==3.1.6
jiter==0.8.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
multidict==6.2.0
numpy==2.2.5
openai==1.72.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
